refactor utils and fix minor bug
